requirements.txt
streamlit>=1.30
python-docx>=1.1.0
